Refactor: Meeting service 패키지 domain으로 이
