+++ src/main/java/com/join/core/meeting/controller/MeetingController.java
@@ -4,9 +4,9 @@ import com.join.core.auth.domain.UserPrincipal;
 import com.join.core.common.response.ApiResponse;
 import com.join.core.meeting.dto.request.MeetingAppendRequest;
 import com.join.core.meeting.dto.response.MeetingResponse;
-import com.join.core.meeting.service.MeetingAppendService;
-import com.join.core.meeting.service.MeetingDeleteService;
-import com.join.core.meeting.service.MeetingReadService;
+import com.join.core.meeting.domain.MeetingAppendService;
+import com.join.core.meeting.domain.MeetingDeleteService;
+import com.join.core.meeting.domain.MeetingReadService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
