+++ src/main/java/com/join/core/meeting/domain/MeetingDeleter.java
@@ -1,0 +1,6 @@
+package com.join.core.meeting.domain;
+
+public interface MeetingDeleter {
+    void delete(Meeting meeting);
+
+}
