+++ src/main/java/com/join/core/meeting/domain/MeetingStore.java
@@ -1,0 +1,6 @@
+package com.join.core.meeting.domain;
+
+public interface MeetingStore {
+    void store(Meeting meeting);
+
+}
