--- src/main/java/com/join/core/meeting/service/MeetingDeleter.java
@@ -1,8 +1,0 @@
-package com.join.core.meeting.service;
-
-import com.join.core.meeting.domain.Meeting;
-
-public interface MeetingDeleter {
-    void delete(Meeting meeting);
-
-}
