--- src/main/java/com/join/core/meeting/service/MeetingStore.java
@@ -1,8 +1,0 @@
-package com.join.core.meeting.service;
-
-import com.join.core.meeting.domain.Meeting;
-
-public interface MeetingStore {
-    void store(Meeting meeting);
-
-}
